+++ js/api.js
@@ -28,4 +28,4 @@ const getData = () => load(Route.GET_DATA, ErrorText.GET_DATA);
 
 const sendData = (body) => load(Route.SEND_DATA, ErrorText.SEND_DATA, Method.POST, body);
 
-export {getData, sendData};
+export { getData, sendData };
